requirements.txt
sqlalchemy==2.0.40
Flask==2.2.5
WTForms==3.1.2
Jinja2==3.1.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
dotenv-cli==3.2.2
click==8.1.7
atomicwrites==1.4.1
joblib==1.3.2
pytest==8.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
pymysql==1.1.0
cryptography==44.0.0
pytz==2024.2
mysql-connector-python==9.2.0
python-dotenv==1.1.0